// management/frontend/src/api/schemas/workspacePublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface WorkspacePublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  description?: string | null;
  logo?: string | null;
  is_active?: boolean;
  creator_name?: string | null;
  updater_name?: string | null;
}