// management/frontend/src/api/zod/system-workspaces/system-workspaces.ts
export const systemWorkspacesReadWorkspacesQuerySkipDefault = 0;
export const systemWorkspacesReadWorkspacesQuerySkipMin = 0;

export const systemWorkspacesReadWorkspacesQueryLimitDefault = 20;
export const systemWorkspacesReadWorkspacesQueryLimitMax = 500;



export const SystemWorkspacesReadWorkspacesQueryParams = zod.object({
  "skip": zod.number().min(systemWorkspacesReadWorkspacesQuerySkipMin).default(systemWorkspacesReadWorkspacesQuerySkipDefault),
  "limit": zod.number().max(systemWorkspacesReadWorkspacesQueryLimitMax).default(systemWorkspacesReadWorkspacesQueryLimitDefault),
  "name": zod.union([zod.string(),zod.null()]).optional(),
  "is_active": zod.union([zod.boolean(),zod.null()]).optional()
})